Player names instead of just amount

+++ DaresGameBot/Game/Data/Card.cs
@@ -13,7 +13,7 @@ internal class Card
     [SheetField(DescriptionTitle)]
     public string Description = null!;
 
-    public virtual bool IsOkayFor(byte playersAmount) => true;
+    public virtual bool IsOkayFor(int playersAmount) => true;
 
     private const string DescriptionTitle = "Текст";
 }
